refactor: change quantity and supplies to uint96 (#412)
**Motivation:**

Prepares the repo for https://github.com/llamaxyz/llama/issues/393#issuecomment-1607967564 to minimize the diff of the core changes

**Modifications:**

Change quantities and supplies from uint128 to uint96. Using uint96 lets us fit both total supply versions and a 64 bit timestamp into a single storage slot, which makes checkpointing total supply cheaper.

**Result:**

Diff on future PRs that implement the new checkpointing logic will be smaller.

+++ script/DeployUtils.sol
@@ -40,7 +40,7 @@ library DeployUtils {
     string comment;
     uint64 expiration;
     address policyholder;
-    uint128 quantity;
+    uint96 quantity;
     uint8 role;
   }
